Checkpoint: Pre-parallelization state
Save current state before beginning Phase 0 parallelization preparation.

Changes include:
- Enhanced plugin system with parser plugins
- Type graph infrastructure
- Code generation improvements (pre.c)
- Standard library enhancements
- Documentation updates (AGENTS.md)
- Various improvements to lexer, parser, and semantic analysis
- Package management updates
- Test infrastructure improvements

This commit establishes a clean baseline before implementing:
- Feature flags for parallel compilation
- Thread-safety annotations
- StrPool mutex support
- Removal of global variables
- Worker thread architecture

See docs/PHASE0-PREPARATION.md for parallelization roadmap.

+++ src/cxy/lang/middle/mangle.c
@@ -28,7 +28,7 @@ static void mangleAstNode(FormatState *state, const AstNode *node, u64 idx)
         else
             format(state,
                    "Lu_{u128}",
-                   (FormatArg[]){{.u64 = node->intLiteral.uValue}});
+                   (FormatArg[]){{.u128 = node->intLiteral.uValue}});
         break;
     case astFloatLit:
         format(state, "Lf_{u64}", (FormatArg[]){{.c = idx}});
